HEX-CyberSphere/core_engine/ai_controller.py
"""
HEX-CyberSphere AI Controller
Python-based AI workflow engine
"""

import json
import numpy as np
from sklearn.ensemble import RandomForestClassifier
import torch
import torch.nn as nn
import pandas as pd
import logging

class AIFramework:
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.models = {}
        self.load_models()
    
    def load_models(self):
        """Load pre-trained AI models"""
        try:
            # Initialize a simple model for demonstration
            self.models['anomaly_detector'] = RandomForestClassifier(n_estimators=100)
            self.logger.info("AI models loaded successfully")
        except Exception as e:
            self.logger.error(f"Failed to load AI models: {e}")
    
    def process_data(self, data):
        """Process data with AI models"""
        try:
            # Convert data to DataFrame
            df = pd.DataFrame(data)
            
            # Perform basic data processing
            processed_data = {
                'rows': len(df),
                'columns': list(df.columns),
                'summary': df.describe().to_dict()
            }
            
            return processed_data
        except Exception as e:
            self.logger.error(f"AI processing failed: {e}")
            return {'error': str(e)}
    
    def detect_anomalies(self, data):
        """Detect anomalies in data"""
        try:
            # Simple anomaly detection logic
            df = pd.DataFrame(data)
            anomalies = []
            
            for column in df.select_dtypes(include=[np.number]).columns:
                mean = df[column].mean()
                std = df[column].std()
                threshold = mean + (2 * std)
                
                column_anomalies = df[df[column] > threshold]
                if not column_anomalies.empty:
                    anomalies.append({
                        'column': column,
                        'anomalies': column_anomalies.to_dict('records')
                    })
            
            return {'anomalies': anomalies}
        except Exception as e:
            self.logger.error(f"Anomaly detection failed: {e}")
            return {'error': str(e)}
    
    def predict_trends(self, data):
        """Predict trends based on historical data"""
        try:
            # Simple trend prediction
            df = pd.DataFrame(data)
            trends = {}
            
            for column in df.select_dtypes(include=[np.number]).columns:
                # Calculate trend direction
                if len(df[column]) > 1:
                    trend = 'increasing' if df[column].iloc[-1] > df[column].iloc[0] else 'decreasing'
                    trends[column] = {
                        'trend': trend,
                        'change': float(df[column].iloc[-1] - df[column].iloc[0])
                    }
            
            return {'trends': trends}
        except Exception as e:
            self.logger.error(f"Trend prediction failed: {e}")
            return {'error': str(e)}

# Example usage
if __name__ == "__main__":
    ai = AIFramework()
    
    # Sample data for testing
    sample_data = {
        'metric1': [1, 2, 3, 4, 5, 100],  # 100 is an anomaly
        'metric2': [10, 20, 30, 40, 50, 60],
        'metric3': [5, 15, 25, 35, 45, 55]
    }
    
    print("Processing data...")
    result = ai.process_data(sample_data)
    print(json.dumps(result, indent=2))
    
    print("\nDetecting anomalies...")
    anomalies = ai.detect_anomalies(sample_data)
    print(json.dumps(anomalies, indent=2))
    
    print("\nPredicting trends...")
    trends = ai.predict_trends(sample_data)
    print(json.dumps(trends, indent=2))