HEX-CyberSphere/core_engine/requirements.txt
flask==2.3.2
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
PyYAML==6.0